requirements.txt
alembic==1.5.8
aniso8601==6.0.0
appdirs==1.4.4
astroid==2.2.5
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
awscli==1.19.111
Babel==2.9.0
bcrypt==3.2.0
botocore==1.20.111
cffi==1.14.5
Click==7.0
colorama==0.4.3
cryptography==3.4.7
distlib==0.3.1
docutils==0.15.2
ecdsa==0.13.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.17.1
greenlet==1.0.0
gunicorn==20.0.4
iniconfig==1.1.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
jwt==1.2.0
lazy-object-proxy==1.4.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.3.1
pycryptodomex==3.10.1
pylint==2.3.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.6.0
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.2.0
python-jose-cryptodome==1.3.2
pytz==2019.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.4.2
six==1.12.0
SQLAlchemy==1.3.4
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.6
virtualenv==20.4.3
Werkzeug==2.0.1
wrapt==1.11.1
WTForms==2.3.3
